+++ kompendium-core/src/main/kotlin/io/bkbn/kompendium/Kontent.kt
@@ -1,13 +1,7 @@
 package io.bkbn.kompendium
 
-import java.util.UUID
-import kotlin.reflect.KClass
-import kotlin.reflect.KType
-import kotlin.reflect.full.isSubclassOf
-import kotlin.reflect.full.memberProperties
-import kotlin.reflect.jvm.javaField
-import kotlin.reflect.typeOf
 import io.bkbn.kompendium.models.meta.SchemaMap
+import io.bkbn.kompendium.models.oas.AnyOfReferencedSchema
 import io.bkbn.kompendium.models.oas.ArraySchema
 import io.bkbn.kompendium.models.oas.DictionarySchema
 import io.bkbn.kompendium.models.oas.EnumSchema
@@ -18,7 +12,16 @@ import io.bkbn.kompendium.models.oas.SimpleSchema
 import io.bkbn.kompendium.util.Helpers.COMPONENT_SLUG
 import io.bkbn.kompendium.util.Helpers.genericNameAdapter
 import io.bkbn.kompendium.util.Helpers.getReferenceSlug
+import io.bkbn.kompendium.util.Helpers.getSimpleSlug
 import io.bkbn.kompendium.util.Helpers.logged
+import java.util.UUID
+import kotlin.reflect.KClass
+import kotlin.reflect.KType
+import kotlin.reflect.full.createType
+import kotlin.reflect.full.isSubclassOf
+import kotlin.reflect.full.memberProperties
+import kotlin.reflect.jvm.javaField
+import kotlin.reflect.typeOf
 import org.slf4j.LoggerFactory
 
 /**
@@ -43,6 +46,19 @@ object Kontent {
   }
 
   /**
+   * Analyzes a [KType] for its top-level and any nested schemas, and adds them to a [SchemaMap], if provided
+   * @param type [KType] to analyze
+   * @param cache Existing schema map to append to
+   * @return an updated schema map containing all type information for [KType] type
+   */
+  fun generateKontent(
+    type: KType,
+    cache: SchemaMap = emptyMap()
+  ): SchemaMap {
+    return generateKTypeKontent(type, cache)
+  }
+
+  /**
    * Analyze a type [T], but filters out the top-level type
    * @param T type to analyze
    * @param cache Existing schema map to append to
@@ -55,6 +71,20 @@ object Kontent {
     val kontentType = typeOf<T>()
     return generateKTypeKontent(kontentType, cache)
       .filterNot { (slug, _) -> slug == (kontentType.classifier as KClass<*>).simpleName }
+  }
+
+  /**
+   * Analyze a type but filters out the top-level type
+   * @param type to analyze
+   * @param cache Existing schema map to append to
+   * @return an updated schema map containing all type information for [T]
+   */
+  fun generateParameterKontent(
+    type: KType,
+    cache: SchemaMap = emptyMap()
+  ): SchemaMap {
+    return generateKTypeKontent(type, cache)
+      .filterNot { (slug, _) -> slug == (type.classifier as KClass<*>).simpleName }
   }
 
   /**
@@ -80,7 +110,7 @@ object Kontent {
         clazz.isSubclassOf(Collection::class) -> handleCollectionType(type, clazz, cache)
         clazz.isSubclassOf(Enum::class) -> handleEnumType(clazz, cache)
         clazz.isSubclassOf(Map::class) -> handleMapType(type, clazz, cache)
-        else -> handleComplexType(clazz, cache)
+        else -> handleComplexType(type, clazz, cache)
       }
     }
   }
@@ -90,32 +120,78 @@ object Kontent {
    * @param clazz Class of the object to analyze
    * @param cache Existing schema map to append to
    */
-  private fun handleComplexType(clazz: KClass<*>, cache: SchemaMap): SchemaMap =
-    when (cache.containsKey(clazz.simpleName)) {
+  private fun handleComplexType(type: KType, clazz: KClass<*>, cache: SchemaMap): SchemaMap {
+    // This needs to be simple because it will be stored under it's appropriate reference component implicitly
+    val slug = type.getSimpleSlug()
+    // Only analyze if component has not already been stored in the cache
+    return when (cache.containsKey(slug)) {
       true -> {
-        logger.debug("Cache already contains ${clazz.simpleName}, returning cache untouched")
+        logger.debug("Cache already contains $slug, returning cache untouched")
         cache
       }
       false -> {
-        logger.debug("${clazz.simpleName} was not found in cache, generating now")
+        logger.debug("$slug was not found in cache, generating now")
         var newCache = cache
+        // Grabs any type parameters as a zip with the corresponding type argument
+        val typeMap = clazz.typeParameters.zip(type.arguments).toMap()
+        // associates each member with a Pair of prop name to property schema
         val fieldMap = clazz.memberProperties.associate { prop ->
           logger.debug("Analyzing $prop in class $clazz")
+          // Grab the field of the current property
           val field = prop.javaField?.type?.kotlin ?: error("Unable to parse field type from $prop")
           logger.debug("Detected field $field")
+          // Yoinks any generic types from the type map should the field be a generic
+          val yoinkBaseType = if (typeMap.containsKey(prop.returnType.classifier)) {
+            logger.debug("Generic type detected")
+            typeMap[prop.returnType.classifier]?.type!!
+          } else {
+            prop.returnType
+          }
+          // converts the base type to a class
+          val yoinkedClassifier = yoinkBaseType.classifier as KClass<*>
+          // in the event of a sealed class, grab all sealed subclasses and create a type from the base args
+          val yoinkedTypes = if (yoinkedClassifier.isSealed) {
+            yoinkedClassifier.sealedSubclasses.map { it.createType(yoinkBaseType.arguments) }
+          } else {
+            listOf(yoinkBaseType)
+          }
+          // if the most up-to-date cache does not contain the content for this field, generate it and add to cache
           if (!newCache.containsKey(field.simpleName)) {
             logger.debug("Cache was missing ${field.simpleName}, adding now")
-            newCache = generateKTypeKontent(prop.returnType, newCache)
+            yoinkedTypes.forEach {
+              newCache = generateKTypeKontent(it, newCache)
+            }
           }
-          val propSchema = ReferencedSchema(field.getReferenceSlug(prop))
+          // TODO This in particular is worthy of a refactor... just not very well written
+          // builds the appropriate property schema based on the property return type
+          val propSchema = if (typeMap.containsKey(prop.returnType.classifier)) {
+            if (yoinkedClassifier.isSealed) {
+              val refs = yoinkedClassifier.sealedSubclasses
+                .map { it.createType(yoinkBaseType.arguments) }
+                .map { ReferencedSchema(it.getReferenceSlug()) }
+              AnyOfReferencedSchema(refs)
+            } else {
+              ReferencedSchema(typeMap[prop.returnType.classifier]?.type!!.getReferenceSlug())
+            }
+          } else {
+            if (yoinkedClassifier.isSealed) {
+              val refs = yoinkedClassifier.sealedSubclasses
+                .map { it.createType(yoinkBaseType.arguments) }
+                .map { ReferencedSchema(it.getReferenceSlug()) }
+              AnyOfReferencedSchema(refs)
+            } else {
+              ReferencedSchema(field.getReferenceSlug(prop))
+            }
+          }
           Pair(prop.name, propSchema)
         }
-        logger.debug("${clazz.simpleName} contains $fieldMap")
+        logger.debug("$slug contains $fieldMap")
         val schema = ObjectSchema(fieldMap)
-        logger.debug("${clazz.simpleName} schema: $schema")
-        newCache.plus(clazz.simpleName!! to schema)
+        logger.debug("$slug schema: $schema")
+        newCache.plus(slug to schema)
       }
     }
+  }
 
   /**
    * Handler for when an [Enum] is encountered