+++ kompendium-core/src/test/kotlin/io/bkbn/kompendium/util/TestModels.kt
@@ -37,6 +37,8 @@ data class TestRequest(
 
 data class TestResponse(val c: String)
 
+data class TestGeneric<T>(val messy: String, val potato: T)
+
 data class TestCreatedResponse(val id: Int, val c: String)
 
 data class ComplexRequest(
@@ -72,3 +74,13 @@ data class OptionalParams(
   @KompendiumParam(ParamType.QUERY) val required: String,
   @KompendiumParam(ParamType.QUERY) val notRequired: String?
 )
+
+sealed class FlibbityGibbit
+
+data class SimpleGibbit(val a: String) : FlibbityGibbit()
+data class ComplexGibbit(val b: String, val c: Int) : FlibbityGibbit()
+
+sealed interface Flibbity<T>
+
+data class Gibbity<T>(val a: T): Flibbity<T>
+data class Bibbity<T>(val b: String, val f: T) : Flibbity<T>